requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.7.0
openai==1.3.0
google-api-python-client==2.108.0
pytube==15.0.0
moviepy==1.0.3
opencv-python==4.8.1.78
pillow==10.1.0
pydub==0.25.1
gtts==2.4.0
elevenlabs==0.2.26
sqlalchemy==2.0.23
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
